move getClassNames into utils
Doing this because I will write a unit test for this function.

There is a performance improvement that I will fix as well, the unit
test would fail if I added it now.

It would also be difficult to see the diff / performance improvement if
I moved the code and did the fix at the same time.